requirements.txt
pillow==7.0.0
pytorch_lightning==0.7.3
tqdm==4.45.0
pyyaml==5.3.1
torch===1.5.0 
torchvision===0.6.0
--find-links https://download.pytorch.org/whl/torch_stable.html